+++ bin/protoc-gen-grpc-ts.js
@@ -26,8 +26,7 @@
 'use strict';
 
 var path = require('path');
-var execSync = require('child_process').execSync;
-var execFile = require('child_process').execFile;
+var spawn = require('child_process').spawn;
 
 var exe_ext = process.platform === 'win32' ? '.exe' : '';
 
@@ -36,15 +35,16 @@ var protoc = path.resolve(__dirname, 'protoc' + exe_ext);
 var plugin = (process.platform === 'win32')
   ? path.resolve(__dirname, 'protoc-gen-ts-plugin.cmd')
   : path.resolve(__dirname, 'protoc-gen-ts-plugin');
-  // : execSync('which protoc-gen-ts-plugin').toString().replace(/\n$/, '').replace(/\r$/, '');
 
 var args = ['--plugin=protoc-gen-ts=' + plugin].concat(process.argv.slice(2));
 
-var child_process = execFile(protoc, args, function(error, stdout, stderr) {
-  if (error) {
-    throw error;
-  }
+var child = spawn(protoc, args, {stdio: 'inherit'});
+
+child.on('error', function(error) {
+  process.stderr.write('failed to run protoc: ' + error.message + '\n');
+  process.exit(1);
 });
 
-child_process.stdout.pipe(process.stdout);
-child_process.stderr.pipe(process.stderr);
+child.on('exit', function(code) {
+  process.exit(code === null ? 1 : code);
+});